app restructured
